helpers/containers/databaseHelper.py
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, Session, declarative_base
from sqlalchemy import Column, Integer, String

from motor.motor_asyncio import AsyncIOMotorClient

from typing import List, Union, Dict, Any, Optional, Tuple, TypeVar, Type, Callable, Awaitable, Annotated, AnyStr, \
    TypedDict

from pydantic import BaseModel
import yaml
from dotenv import load_dotenv