pyproject.toml
[tool.poetry]
name = "blacklotuscontainer001"
version = "0.1.0"
description = ""
authors = ["Shane B <shaneb@datanamix.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py-cord = {extras = ["speed"], version = "^2.6.0"}
dependency-injector = "^4.41.0"
sqlalchemy = "^2.0.32"
motor = "^3.5.1"
pyyaml = "^6.0.2"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
rich = "^13.8.0"
art = "^6.2"
arrow = "^1.3.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
